Training/requirements.txt
cellpose>=4.0
numpy>=1.24
scipy>=1.10
scikit-image>=0.21
tifffile>=2023.7
nd2>=0.10
bioimageio.core>=0.6
pyyaml>=6.0
matplotlib>=3.7
torch>=2.0
tqdm>=4.65
